+++ src/main/java/pl/lcc/yasmart/common/account/AccountRepository.java
@@ -1,6 +1,7 @@
-package pl.lcc.yasmart.common;
+package pl.lcc.yasmart.common.account;
 
 import org.springframework.data.repository.CrudRepository;
+import pl.lcc.yasmart.common.account.Account;
 
 public interface AccountRepository extends CrudRepository<Account, Long> {
     Account findByName(String name);
