+++ src/main/java/pl/lcc/yasmart/common/tag/TagRepository.java
@@ -1,6 +1,7 @@
-package pl.lcc.yasmart.common;
+package pl.lcc.yasmart.common.tag;
 
 import org.springframework.data.repository.CrudRepository;
+import pl.lcc.yasmart.common.tag.Tag;
 
 import java.util.UUID;
